+++ src/components/Toast/useToast.tsx
@@ -1,16 +1,19 @@
 import { useToastContext } from './useToastContext';
 import { PrimitiveToastProps as ToastProps } from './Toast';
-import { ReactNode } from 'react';
+import { ReactNode, useCallback } from 'react';
 
 export const useToast = () => {
   const { addToast } = useToastContext();
-  const toast = (options: ToastProps | ReactNode) => {
-    if (typeof options === 'object' && options !== null && 'title' in options) {
-      addToast(options);
-    } else {
-      addToast({ title: options });
-    }
-  };
+  const toast = useCallback(
+    (options: ToastProps | ReactNode) => {
+      if (typeof options === 'object' && options !== null && 'title' in options) {
+        addToast(options);
+      } else {
+        addToast({ title: options });
+      }
+    },
+    [addToast],
+  );
 
   return toast;
 };