+++ src/patterns/SaleHeaderBanner/SaleHeaderBanner.tsx
@@ -18,6 +18,8 @@ export interface SaleHeaderBannerProps extends ComponentProps<'div'> {
    * What is the title of the auction?
    */
   auctionTitle: React.ReactNode;
+  /** Text to be displayed as a badge. */
+  badgeText?: string;
   /**
    * The URL of the banner image
    */
@@ -102,6 +104,7 @@ const SaleHeaderBanner = forwardRef<HTMLDivElement, SaleHeaderBannerProps>(
     {
       auctionEndTime,
       auctionTitle,
+      badgeText,
       imageSrcUrl,
       imageSrcSet,
       imageSizes,
@@ -161,6 +164,11 @@ const SaleHeaderBanner = forwardRef<HTMLDivElement, SaleHeaderBannerProps>(
             ) : null}
             <Text variant={TextVariants.badge}>{headerLabel}</Text>
             <Text variant={TextVariants.title1}>{auctionTitle}</Text>
+            {badgeText && (
+              <Text variant={TextVariants.badge} className={`${baseClassName}__badge`}>
+                {badgeText}
+              </Text>
+            )}
             <Text variant={TextVariants.string2} className={`${baseClassName}__location`}>
               {location}
             </Text>