+++ CHANGELOG.md
@@ -1,3 +1,57 @@
+## [1.182.1](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.182.0...v1.182.1) (2025-08-21)
+
+
+### Bug Fixes
+
+* **NotificationBanner:** L3-7547 Add Notification Banner child style ([#700](https://github.com/PhillipsAuctionHouse/seldon/issues/700)) ([1680bd6](https://github.com/PhillipsAuctionHouse/seldon/commit/1680bd6b03016c4a394c5bf9d4fc06ebb6e5a6eb))
+
+# [1.182.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.181.0...v1.182.0) (2025-08-21)
+
+
+### Features
+
+* **header:** L3-5714 Small slice of a button on narrow windows ([#689](https://github.com/PhillipsAuctionHouse/seldon/issues/689)) ([3d66b67](https://github.com/PhillipsAuctionHouse/seldon/commit/3d66b6736dd038349a10f48bea41d6e214be2c45))
+
+
+### Reverts
+
+* **_type:** L3-7279 typography tokens ([#698](https://github.com/PhillipsAuctionHouse/seldon/issues/698)) ([9a942b1](https://github.com/PhillipsAuctionHouse/seldon/commit/9a942b1b99901b61b02425f2664f222c092248ba))
+
+# [1.181.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.180.0...v1.181.0) (2025-08-21)
+
+
+### Features
+
+* **HeroBanner:** L3-7300 add Text Banner ([#674](https://github.com/PhillipsAuctionHouse/seldon/issues/674)) ([066f8d5](https://github.com/PhillipsAuctionHouse/seldon/commit/066f8d5fd81a02a9417ed3b4f02dfdfea712c589))
+
+# [1.180.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.179.0...v1.180.0) (2025-08-20)
+
+
+### Features
+
+* **_type:** L3-7279 typography tokens ([#687](https://github.com/PhillipsAuctionHouse/seldon/issues/687)) ([020e114](https://github.com/PhillipsAuctionHouse/seldon/commit/020e11461a6c9704c566839df78bee4caa132406))
+
+# [1.179.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.178.0...v1.179.0) (2025-08-20)
+
+
+### Features
+
+* **SaleHeaderBanner:** L3-7403 add badgeText prop to SaleHeaderBanner ([#693](https://github.com/PhillipsAuctionHouse/seldon/issues/693)) ([140881c](https://github.com/PhillipsAuctionHouse/seldon/commit/140881c11c151956ae7e7cdf709362aa37aa780f))
+
+# [1.178.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.177.0...v1.178.0) (2025-08-19)
+
+
+### Features
+
+* **AnnouncementBanner:** L3-7452 Create Announcement Bar ([#681](https://github.com/PhillipsAuctionHouse/seldon/issues/681)) ([f5dccb5](https://github.com/PhillipsAuctionHouse/seldon/commit/f5dccb59d44c6515f63c8640c14ebe559696d4f9))
+
+# [1.177.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.176.1...v1.177.0) (2025-08-19)
+
+
+### Features
+
+* **combobox:** L3-6813 allow specification of minimum chars before opening ([#690](https://github.com/PhillipsAuctionHouse/seldon/issues/690)) ([184c3ae](https://github.com/PhillipsAuctionHouse/seldon/commit/184c3aef7dfc9aacc128469016ab78de2b3d7e3a))
+
 ## [1.176.1](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.176.0...v1.176.1) (2025-08-19)